Semntic fix. Comment out duplicate code
Code is at the moment duplicated. Could be used later if a change to
transaction entitiy is considered.
At the moment the project id has to be provided upon creation. Therefore
setting it during payment is not necessary.
If project_id is marked nullable for a transaction it might become viable to
set it during payment instead during creation.