Project ID is set during Payment
The Project ID is connected with the payment type. Depending on where the payment is handled bank vs paypal etc.
a different Project ID is set. The Project ID is a part of Giropay`s processing system.

Creation only needs amount, purpose and currency. The Project ID is set during GiroCheckout::Transaction#pay.